output consistency: rework input data to avoid circular dependency
