+++ misc/python/materialize/output_consistency/execution/query_execution_manager.py
@@ -67,7 +67,7 @@ class QueryExecutionManager:
                 f"Setup for evaluation strategy '{strategy.name}'"
             )
             executor = self.executors.get_executor(strategy)
-            ddl_statements = strategy.generate_sources(input_data)
+            ddl_statements = strategy.generate_sources(input_data.types_input)
 
             for sql_statement in ddl_statements:
                 self.output_printer.print_sql(sql_statement)
